Move SigningCert to own package
As this will be shared between root CA generation, and the signing code,
move it into its own module.  The API will be enhanced in upcoming
changes.
